func(V1): bm25

+++ tcvectordb/src/main/java/com/tencent/tcvectordb/utils/ConvertUtils.java
@@ -6,9 +6,9 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class ConvertUtils {
-    public static List<List<Object>> convertPairToList(List<Pair<Long, Double>> pairList) {
+    public static List<List<Object>> convertPairToList(List<Pair<Long, Float>> pairList) {
         List<List<Object>> result = new ArrayList<>();
-        for (Pair<Long, Double> p : pairList) {
+        for (Pair<Long, Float> p : pairList) {
             List<Object> list = new ArrayList<>();
             list.add(p.getLeft());
             list.add(p.getRight());
